intelligence-backend/requirements.txt
flask
gunicorn
flask-socketio
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
celery
redis
requests
python-dotenv
google-generativeai